Linux compilation fixes

+++ Visualization_Library_SDK/src/vl/vl3DS.cpp
@@ -774,7 +774,7 @@ ref<ResourceDatabase> vl::load3DS(VirtualFile* file)
       ref<DrawElementsUInt> polys = new DrawElementsUInt;
       geom->primitives()->push_back( polys.get() );
       polys->indices()->resize( index_buffer.size() );
-      VL_CHECK( polys->indices()->bytesUsed() == (int)sizeof(index_buffer[0]) * (int)index_buffer.size() )
+      VL_CHECK( polys->indices()->bytesUsed() == sizeof(index_buffer[0]) * index_buffer.size() )
       memcpy(polys->indices()->ptr(), &index_buffer[0], polys->indices()->bytesUsed());
 
       // matrix
